% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmr.R
\name{constroi_dataframe}
\alias{constroi_dataframe}
\title{Função que constroi o dataframe modelo utilizado para as análises realizadas pela plataforma "Quem me representa?"}
\usage{
constroi_dataframe(proposicao, votacao, votos)
}
\arguments{
\item{proposicao}{Uma proposição especifica recuperada pelo método fetch_proposicao()}

\item{votacao}{Uma votação específica recuperada pelo método fetch_votacao()}

\item{votos}{Os votos referentes a esta votação em questão}
}
\value{
Dataframe contendo 10 colunas com as informações: Nome do parlamentar, ID do parlamentar,
   sigla do partido, sigla da uf, voto, número da proposição, ano, ementa, horário da votação e
   orientação do governo
}
\description{
Função que constroi o dataframe modelo utilizado para as análises realizadas pela plataforma "Quem me representa?"
}
\examples{
pec241 <- fecth_proposicao(2088351)
votacao_segundoturno_pec241 <- fetch_votacao(7252)
votos_segundoturno_pec241 <- fetch_votos(7252)
dataframe_pec241 <- constroi_dataframe(pec241, votacao_segundoturno_pec241, votos_segundoturno_pec241)

}
