% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.assert_dataframe_completo}
\alias{.assert_dataframe_completo}
\title{Garantees that the dataframe x has all the columns passed by y.}
\usage{
.assert_dataframe_completo(x, y)
}
\arguments{
\item{x}{dataframe}

\item{y}{vector of characters containing the names of columns.}
}
\description{
Garantees that the dataframe x has all the columns passed by y.
}
