% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.camara_api}
\alias{.camara_api}
\title{Wraps an access to the camara API given a relative path and query arguments.}
\usage{
.camara_api(path = NULL, query = NULL, asList = FALSE)
}
\arguments{
\item{path}{URL relative to the API base URL}

\item{query}{Query parameters}
}
\description{
Wraps an access to the camara API given a relative path and query arguments.
}
