% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.coerce_types}
\alias{.coerce_types}
\title{Garantees that the dataframe obj has all the correct types passed by types.}
\usage{
.coerce_types(obj, types)
}
\arguments{
\item{obj}{dataframe}

\item{types}{named vector of the columns names and types}
}
\description{
Garantees that the dataframe obj has all the correct types passed by types.
}
