% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.fetch_itens}
\alias{.fetch_itens}
\title{Abstracts the pagination logic. There are three situations on a request from the API:
1. The items number is less than the max by request, that is 100.
2. The items number is divisible by 100.
3. The items number is not divisible by 100.}
\usage{
.fetch_itens(query, API_path)
}
\arguments{
\item{query}{query parameters}

\item{API_path}{API path}
}
\description{
Case 1 and 2 are solved using the same logic: Fetches from the API the exact quantity
to fill n pages. 100 items fill into 1 page, 200 items fill into 2 pages and so on...
}
\details{
Case 3 is solved using the previous thinking adding a little detail: Fetches all the items
until it fills completly the pages with 100 items each one, then insert the remaining
items. 530 items can also be read as 500 items + 30 items, then 5 pages with 100 items and
1 page with 30 items.
}
