% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatorias.R
\name{.fetch_relatorias_camara}
\alias{.fetch_relatorias_camara}
\title{Fetches proposição's relatoria historic from Camara}
\usage{
.fetch_relatorias_camara(proposicao_id)
}
\arguments{
\item{proposicao_id}{Proposição's ID from Camara}
}
\value{
Dataframe containing detailed information about relatorias of a proposição from Camara
}
\description{
Returns a dataframe containing the relatoria's historic, with date and relator's name from Camara.
}
