% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatorias.R
\name{.fetch_relatorias_senado}
\alias{.fetch_relatorias_senado}
\title{Fetches proposição's relatoria historic from Senado}
\usage{
.fetch_relatorias_senado(proposicao_id)
}
\arguments{
\item{proposicao_id}{Proposição's ID from Senado}
}
\value{
Dataframe containing detailed information about relatorias of a proposição from Senado
}
\description{
Returns a dataframe containing the relatoria's historic, with date and relator's name from Senado.
}
