% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.fetch_using_queries}
\alias{.fetch_using_queries}
\title{Fetches a proposition using a list of queries}
\usage{
.fetch_using_queries(parametros, API_path, asList = FALSE)
}
\arguments{
\item{parametros}{queries used on the search}
}
\value{
Dataframe containing information about the proposition.
}
\description{
Fetches a proposition using a list of queries
}
\examples{
pec241 <- .fetch_using_queries(siglaTipo = "PEC", numero = 241, ano = 2016)

}
