% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_dataframe}
\alias{.get_dataframe}
\title{In case of receiving a list, this function converts the list into a dataframe.}
\usage{
.get_dataframe(x)
}
\arguments{
\item{x}{List}
}
\description{
In case of receiving a list, this function converts the list into a dataframe.
}
