% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_json}
\alias{.get_json}
\title{Recovers a json from a URL using HTTP.}
\usage{
.get_json(response)
}
\arguments{
\item{full_link}{URL admitting GET HTTP requests.}
}
\value{
the json.
}
\description{
Recovers a json from a URL using HTTP.
}
\examples{
pec241_json <- .get_json("https://dadosabertos.camara.leg.br/api/v2/proposicoes/2088351")
}
