% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{.HTTP_CACHE}
\alias{.HTTP_CACHE}
\title{Used to store HTTP responses.}
\format{An object of class \code{NULL} of length 0.}
\usage{
.HTTP_CACHE
}
\description{
Used to store HTTP responses.
}
\keyword{datasets}
