% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.print_warning_and_list}
\alias{.print_warning_and_list}
\title{Prints a warning and a list.}
\usage{
.print_warning_and_list(msg, l)
}
\arguments{
\item{msg}{warning message}

\item{l}{list}
}
\description{
Prints a warning and a list.
}
