% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.put_in_cache}
\alias{.put_in_cache}
\title{Stores a value in the cache.}
\usage{
.put_in_cache(key, value)
}
\arguments{
\item{key}{Key to store}

\item{value}{Value to store}
}
\description{
Stores a value in the cache.
}
