% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatorias.R
\name{.rename_relatorias_senado_columns}
\alias{.rename_relatorias_senado_columns}
\title{Renames columns from Senado Relatoria's dataframe}
\usage{
.rename_relatorias_senado_columns(df)
}
\arguments{
\item{df}{Senado Relatoria's dataframe}
}
\value{
Senado Relatoria's dataframe renamed
}
\description{
Replaces spaces to underscore and format to lowercase
}
