% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.rename_tramitacao_df}
\alias{.rename_tramitacao_df}
\title{Renomeia as colunas do dataframe de movimentação no Senado}
\usage{
.rename_tramitacao_df(df)
}
\arguments{
\item{df}{Dataframe da votação no Senado}
}
\value{
Dataframe com as colunas renomeadas
}
\description{
Renomeia as colunas do dataframe de movimentação no Senado usando o padrão
de underscore e letras minúsculas
}
