% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.senado_api}
\alias{.senado_api}
\title{Wraps an access to the senate API given a relative path and query arguments.}
\usage{
.senado_api(path = NULL, query = NULL, asList = FALSE)
}
\arguments{
\item{path}{URL relative to the API base URL}

\item{query}{Query parameters}

\item{asList}{If return should be a list or a dataframe}
}
\description{
Wraps an access to the senate API given a relative path and query arguments.
}
