% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.switch_types}
\alias{.switch_types}
\title{Returns a conversion function given a type name.}
\usage{
.switch_types(x)
}
\arguments{
\item{x}{type name}
}
\description{
Returns a conversion function given a type name.
}
