% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.to_tibble}
\alias{.to_tibble}
\title{Converts a vector of integer into a tibble. Also useful when the user is working with
dplyr functions.}
\usage{
.to_tibble(num)
}
\arguments{
\item{num}{A integer vector}
}
\value{
A tibble
}
\description{
Converts a vector of integer into a tibble. Also useful when the user is working with
dplyr functions.
}
\examples{
df <- rcongresso:::.to_tibble(c(1,2))
}
