% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.to_underscore}
\alias{.to_underscore}
\title{Renames a vector with the pattern of underscores and lowercases}
\usage{
.to_underscore(x)
}
\arguments{
\item{x}{Strings vector}
}
\value{
Vector containing the renamed strings.
}
\description{
Renames each item from vector with the pattern: split by underscore and lowercase
}
