% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.verifica_id}
\alias{.verifica_id}
\title{Check if a proposition/party exists and returns a warning message when it does not.}
\usage{
.verifica_id(x, message)
}
\arguments{
\item{x}{returned object from proposition or party function}

\item{message}{error message expected}
}
\description{
Check if a proposition/party exists and returns a warning message when it does not.
}
