% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.verifica_parametros_entrada}
\alias{.verifica_parametros_entrada}
\title{Verifies from the input if all the parameters are available and handles correctly
about the transformation into a valid URL query.}
\usage{
.verifica_parametros_entrada(parametros)
}
\arguments{
\item{parametros}{A list of parameters from the input}
}
\value{
A list of parameters without NULL
}
\description{
Verifies from the input if all the parameters are available and handles correctly
about the transformation into a valid URL query.
}
\examples{
parametros <- .verifica_parametros_entrada(list(NULL, itens=100, pagina=1))
}
