% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deputados.R
\name{extract_partido_estado_autor}
\alias{extract_partido_estado_autor}
\title{Get the state and party of an author}
\usage{
extract_partido_estado_autor(uri)
}
\arguments{
\item{uri}{uri that contains data about the author}
}
\value{
State and party
}
\description{
Return state and party
}
