% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgaos.R
\name{fetch_agenda_orgaos_camara}
\alias{fetch_agenda_orgaos_camara}
\title{Fetches all Chamber of deputies organs schedule}
\usage{
fetch_agenda_orgaos_camara(initial_date, end_date)
}
\arguments{
\item{initial_date}{initial date dd/mm/yyyy}

\item{end_date}{end_date dd/mm/yyyy}
}
\value{
Returns a dataframe
}
\description{
Return a dataframe containing information about the Chamber of deputies organs schedule
}
\examples{
fetch_agenda_orgaos_camara('12/05/2018', '26/05/2018')
}
