% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agenda.R
\name{fetch_agenda_senado}
\alias{fetch_agenda_senado}
\title{Get the Senate's schedule}
\usage{
fetch_agenda_senado(initial_date)
}
\arguments{
\item{initial_date}{inital date yyyy-mm-dd}
}
\value{
list
}
\description{
Return a list with 3 dataframes: schedule, bills and speakers. All the dfs contains a column named
codigo_sessao
}
\examples{
fetch_agenda_senado('2018-07-03')
}
