% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agenda.R
\name{fetch_agenda_senado_comissoes}
\alias{fetch_agenda_senado_comissoes}
\title{Comissions schedule Senate}
\usage{
fetch_agenda_senado_comissoes(initial_date, end_date)
}
\arguments{
\item{initial_date}{initial date yyyy-mm-dd}

\item{end_date}{end date yyyy-mm-dd}
}
\value{
Dataframe
}
\description{
Return a dataframe with the Senate's Comissions schedule
}
\examples{
\dontrun{
fetch_agenda_senado_comissoes('2016-05-15', '2016-05-25')
}
}
