% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgaos.R
\name{fetch_agendas_comissoes_camara_auxiliar}
\alias{fetch_agendas_comissoes_camara_auxiliar}
\title{Baixa dados da agenda de um orgão da Camara}
\usage{
fetch_agendas_comissoes_camara_auxiliar(orgao_id, initial_date, end_date)
}
\arguments{
\item{orgao_id}{ID do orgão}

\item{initial_date}{data inicial no formato dd/mm/yyyy}

\item{end_date}{data final no formato dd/mm/yyyy}
}
\value{
Dataframe
}
\description{
Retorna um dataframe contendo dados sobre a agenda de um orgão da camara
}
\examples{
fetch_agendas_comissoes_camara_auxiliar('2003', '12/05/2018', '26/05/2018')
}
