% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deputados.R
\name{fetch_all_deputados}
\alias{fetch_all_deputados}
\title{Retrieves details about all deputy}
\usage{
fetch_all_deputados()
}
\value{
Dataframe containing details about all deputys.
}
\description{
ID, name, birth date, birth city among other informations are returned.
}
\examples{
deps <- fetch_all_deputados()
}
