% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senadores.R
\name{fetch_all_senadores}
\alias{fetch_all_senadores}
\title{Fetches details abaout all senators}
\usage{
fetch_all_senadores(ids_sen)
}
\arguments{
\item{ids_dep}{Dataframe containing all senators IDs}
}
\value{
Dataframe containing details about the senators
}
\description{
Fetches details about senators from a given list
}
