% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposicoes.R
\name{fetch_apensadas_camara}
\alias{fetch_apensadas_camara}
\title{Fetch the propositions appended to a proposition in the Camara}
\usage{
fetch_apensadas_camara(prop_id)
}
\arguments{
\item{prop_id}{Proposition's ID}
}
\value{
A vector of characters containing the ids of the appended propositions
}
\description{
Returns a vector containing the ids of the appended propositions
}
\examples{
fetch_apensadas(2121442)
}
