% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposicoes.R
\name{fetch_autor_camara}
\alias{fetch_autor_camara}
\title{Fetches proposition's author}
\usage{
fetch_autor_camara(proposicao_id = NULL)
}
\arguments{
\item{proposicao_id}{Proposition's ID}
}
\value{
A dataframe containing details about the author of the proposition
}
\description{
Fetches a dataframe containing information about the author of the proposition
}
\examples{
fetch_autor_camara(2121442)
}
