% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposicoes.R
\name{fetch_autores}
\alias{fetch_autores}
\title{Fetches proposition's authors}
\usage{
fetch_autores(proposicao_id = NULL, casa)
}
\arguments{
\item{proposicao_id}{Proposition's ID}
}
\value{
A dataframe containing the basic information about the authors of the proposition
}
\description{
Fetches a dataframe containing basic information about the authors of the proposition
}
\examples{
\dontrun{
.fetch_autores(2121442, 'camara')
.fetch_autores(91341, 'senado')
}
}
