% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposicoes.R
\name{fetch_autores_camara}
\alias{fetch_autores_camara}
\title{Fetches proposition's authors}
\usage{
fetch_autores_camara(proposicao_id = NULL, sigla_tipo = "")
}
\arguments{
\item{proposicao_id}{Proposition's ID}

\item{sigla_tipo}{Proposition's initials}
}
\value{
A dataframe containing the basic information about the authors of the proposition
}
\description{
Fetches a dataframe containing basic information about the authors of the proposition
}
\examples{
\dontrun{
fetch_autores_camara(2121442)
}
}
