% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comissoes.R
\name{fetch_comissao_camara}
\alias{fetch_comissao_camara}
\title{Fetches camara comissao}
\usage{
fetch_comissao_camara(sigla = NULL)
}
\arguments{
\item{sigla}{Comissao's id}
}
\value{
Returns a dataframe containing information about the comissao requested
}
\description{
Fetches a dataframe containing information about the comissao requested
}
