% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comissoes.R
\name{fetch_composicao_comissoes_senado}
\alias{fetch_composicao_comissoes_senado}
\title{Fetches comissions on Senate}
\usage{
fetch_composicao_comissoes_senado(sigla)
}
\arguments{
\item{sigla}{Comissao's id}
}
\value{
Returns a dataframe containing information about the comission requested
}
\description{
Fetches a dataframe containing information about the comission requested
}
