% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposicoes.R
\name{fetch_deferimento}
\alias{fetch_deferimento}
\title{Requirements's deferments}
\usage{
fetch_deferimento(proposicao_id)
}
\arguments{
\item{proposicao_id}{ID of requirements}
}
\value{
Dataframe
}
\description{
Verify deferments to a id's list of requirements
}
\examples{
fetch_deferimento(c("102343", "109173", "115853"))
}
