% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deputados.R
\name{fetch_deputado}
\alias{fetch_deputado}
\title{Recupera da API os detalhes de um deputado.}
\usage{
fetch_deputado(dep_id)
}
\arguments{
\item{dep_id}{ID do Deputado}
}
\value{
Lista contendo detalhes sobre o deputado.
}
\description{
Recupera da API os detalhes de um deputado.
}
\examples{
abel_mesquita_info <- fetch_deputado(178957)

}
