% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deputados.R
\name{fetch_despesas_deputado}
\alias{fetch_despesas_deputado}
\title{Recupera todos os gastos da cota parlamentar, dos últimos seis meses, um deputado específico.}
\usage{
fetch_despesas_deputado(dep_id)
}
\arguments{
\item{dep_id}{ID do deputado}
}
\value{
Dataframe contendo detalhes sobre os gastos do deputado
}
\description{
Recupera todos os gastos da cota parlamentar, dos últimos seis meses, um deputado específico.
}
\examples{
gastos_abel_mesquita <- fetch_despesas_deputado(178957)

}
