% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emendas.R
\name{fetch_emendas_senado}
\alias{fetch_emendas_senado}
\title{Returns emendas of a proposição from Senado}
\usage{
fetch_emendas_senado(bill_id)
}
\arguments{
\item{bill_id}{Proposição's ID from senado.}
}
\value{
Dataframe with informations about emendas of a proposição from Senado.
}
\description{
Fetchs a dataframe with emendas's data of a proposição from Senado.
}
