% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requerimentos.R
\name{fetch_events_requerimento_camara}
\alias{fetch_events_requerimento_camara}
\title{Fetch events of a requerimento}
\usage{
fetch_events_requerimento_camara(req_id)
}
\arguments{
\item{req_id}{ID of a requerimento}
}
\value{
Dataframe
}
\description{
Returns a dataframe with events of a given requerimento (presentation, deferral, etc.)
}
