% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frentes.R
\name{fetch_frentes}
\alias{fetch_frentes}
\title{Fetches info about frente}
\usage{
fetch_frentes(id_frente, casa)
}
\arguments{
\item{id_frente}{Frente ID}
}
\value{
A dataframe containing details about the frente
}
\description{
Fetches a dataframe containing information about the frentes
}
\examples{
fetch_frentes(54012, "camara")
}
