% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frentes.R
\name{fetch_frentes_camara}
\alias{fetch_frentes_camara}
\title{Fetches info about frente}
\usage{
fetch_frentes_camara(id_frente)
}
\arguments{
\item{id_frente}{Frente ID}
}
\value{
A dataframe containing details about the frente
}
\description{
Fetches a dataframe containing information about the frentes of the Câmara
}
\examples{
fetch_frentes_camara(54012)
}
