% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deputados.R
\name{fetch_frentes_deputado}
\alias{fetch_frentes_deputado}
\title{Fetches all fronts from deputy}
\usage{
fetch_frentes_deputado(id_dep, legislatura_base = .LEGISLATURA_ATUAL)
}
\arguments{
\item{id_dep}{Deputy's ID}

\item{legislatura_base}{Fronts legislature}
}
\value{
Dataframe containing all fronts from a deputy
}
\description{
Fetches all fronts from a deputy from the given legislature
}
