% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partidos.R
\name{fetch_id_partido}
\alias{fetch_id_partido}
\title{Retrieves the party's ID from its abbreviation.}
\usage{
fetch_id_partido(sigla)
}
\arguments{
\item{sigla}{Party's abbreviation}
}
\value{
Party's ID.
}
\description{
The function can be used to fetch a vector of ids as well, in case of many parties.
}
\examples{
p <- fetch_id_partido(c("PT","PSDB","PP","PMDB"))
}
\seealso{
\code{\link[rcongresso]{fetch_id_proposicao}}
}
