% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposicoes.R
\name{fetch_id_proposicao}
\alias{fetch_id_proposicao}
\title{Recupera da API o ID da preposição de acordo com o tipo, número da proposição e ano que
esta foi feita.}
\usage{
fetch_id_proposicao(tipo, numero, ano)
}
\arguments{
\item{tipo}{Tipo da proposição (i.e., PEC, PL, PDC)}

\item{numero}{Número da proposição}

\item{ano}{Ano da proposição}
}
\value{
ID da proposição
}
\description{
Recupera da API o ID da preposição de acordo com o tipo, número da proposição e ano que
esta foi feita.
}
\examples{
pec241_id <- fetch_id_proposicao("PEC", 241, 2016)

}
