% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deputados.R
\name{fetch_ids_deputados_by_leg}
\alias{fetch_ids_deputados_by_leg}
\title{Fetches all deputys IDsfrom a given legislature}
\usage{
fetch_ids_deputados_by_leg(legislatura = .LEGISLATURA_ATUAL)
}
\arguments{
\item{legislatura_id}{Legislature's id}
}
\value{
Dataframe containing all deputys IDs
}
\description{
Fetches all deputys IDs from a given legislature
}
