% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposicoes.R
\name{fetch_ids_relacionadas}
\alias{fetch_ids_relacionadas}
\title{Fetches all propositions'ids related to a proposition}
\usage{
fetch_ids_relacionadas(id, casa)
}
\arguments{
\item{id}{Proposition's ID}

\item{casa}{Câmara or sendo}
}
\value{
Dataframe containing all the related propositions'ids.
}
\description{
Returns all propositions'ids related to a proposition.
}
