% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senadores.R
\name{fetch_ids_senadores}
\alias{fetch_ids_senadores}
\title{Fetches senators id}
\usage{
fetch_ids_senadores(legis_initial, legis_final)
}
\arguments{
\item{legis_initial}{Initial legisture}

\item{legis_final}{Final legisture}
}
\value{
Dataframe containing the senators id.
}
\description{
Returns its id.
}
\examples{
senators <- fetch_ids_senadores(40, 56)
}
