% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senadores.R
\name{fetch_ids_senadores_em_exercicio}
\alias{fetch_ids_senadores_em_exercicio}
\title{Fetches senators in office ids}
\usage{
fetch_ids_senadores_em_exercicio()
}
\value{
Dataframe containing if of the senators in office
}
\description{
Fetches the ids of the senators in office
}
