% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgaos.R
\name{fetch_orgao_camara}
\alias{fetch_orgao_camara}
\title{Fetches Câmara órgão}
\usage{
fetch_orgao_camara(sigla = NULL)
}
\arguments{
\item{sigla}{Órgão's id}
}
\value{
Returns a dataframe containing information about the órgão requested
}
\description{
Fetches a dataframe containing information about the órgão requested
}
