% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgaos.R
\name{fetch_orgaos_camara}
\alias{fetch_orgaos_camara}
\title{Baixa os órgãos na Câmara}
\usage{
fetch_orgaos_camara()
}
\value{
Dataframe contendo os órgãos da Câmara
}
\description{
Retorna um dataframe contendo os órgãos da câmara
}
