% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votacoes.R
\name{fetch_orientacoes}
\alias{fetch_orientacoes}
\title{Fetch how the groups and parties in the chamber of deputies
instructed their members to vote on a given voting.}
\usage{
fetch_orientacoes(id_votacao)
}
\arguments{
\item{id_votacao}{Voting's ID.}
}
\value{
Dataframe containing the parties and their votes
}
\description{
Fetch how the groups and parties in the chamber of deputies
instructed their members to vote on a given voting.
}
\examples{
orientacoesbancada_segundoturno_pec241 <- fetch_orientacoes(7252)

}
