% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agenda.R
\name{fetch_pauta_camara}
\alias{fetch_pauta_camara}
\title{Get the agenda of a meeting}
\usage{
fetch_pauta_camara(id, hora_inicio, hora_fim, sigla_orgao, nome_orgao)
}
\arguments{
\item{id}{event id}

\item{hora_inicio}{inital time}

\item{hora_fim}{end time}

\item{sigla_orgao}{Acronym of the organ}

\item{nome_orgao}{Name of the organ}
}
\value{
Dataframe
}
\description{
Return a dataframe with data about the agenda
}
\examples{
\dontrun{
fetch_pauta_camara('53184', '2018-07-03T10:00', '2018-07-03T12:37', 'CVT', 'Comissão de Viação e Transportes VIAÇÃO E TRANSPORTES')
}
}
