% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposicoes.R
\name{fetch_proposicao}
\alias{fetch_proposicao}
\title{Recupera da API uma proposição contendo detalhes adicionais.}
\usage{
fetch_proposicao(id_props)
}
\arguments{
\item{id_prop}{ID da proposição}
}
\value{
Lista contendo informações sobre a proposição
}
\description{
Recupera da API uma proposição contendo detalhes adicionais.
}
\examples{
pec241 <- fetch_proposicao(2088351)

}
