% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votacoes.R
\name{fetch_proposicao_from_votacao}
\alias{fetch_proposicao_from_votacao}
\title{Dado o id da votação, retorna a proposição associada àquela votação.}
\usage{
fetch_proposicao_from_votacao(id_votacao)
}
\arguments{
\item{id_votacao}{ID da votação}
}
\value{
Dataframe contendo informações sobre a proposição relacionada à votação
}
\description{
Dado o id da votação, retorna a proposição associada àquela votação.
}
\examples{
pec241 <- fetch_proposicao_from_votacao(7252)

}
