% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposicoes.R
\name{fetch_proposicao_por_id}
\alias{fetch_proposicao_por_id}
\title{Fetches details from a proposition.}
\usage{
fetch_proposicao_por_id(id)
}
\arguments{
\item{id}{Proposition's ID}
}
\value{
Dataframe containing information about the proposition.
}
\description{
Fetches details from a proposition.
}
\examples{
pec241 <- fetch_proposicao_por_id(2088351)

}
