% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposicoes.R
\name{fetch_proposicao_senado}
\alias{fetch_proposicao_senado}
\title{Fetches a proposition in the Senate}
\usage{
fetch_proposicao_senado(id = NULL)
}
\arguments{
\item{id}{Proposition's ID}
}
\value{
Dataframe containing all the info about the proposition;
}
\description{
Returns the proposition info
}
\examples{
prop_pls229 <- fetch_proposicao_senado(91341)
}
