% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposicoes.R
\name{fetch_proposicao_senado_sigla}
\alias{fetch_proposicao_senado_sigla}
\title{Fetches a proposition in the Senate}
\usage{
fetch_proposicao_senado_sigla(sigla, numero, ano)
}
\arguments{
\item{sigla}{Proposition's sigla}

\item{numero}{Proposition's numero}

\item{ano}{Proposition's ano}
}
\value{
Dataframe containing all the info about the proposition;
}
\description{
Returns the proposition info
}
\examples{
prop_pls229 <- fetch_proposicao_senado(pls, 229, 2009)
}
