% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposicoes.R
\name{fetch_proposicoes}
\alias{fetch_proposicoes}
\title{Recupera da API todas as proposições disponíveis na base de dados da Câmara dos Deputados.}
\usage{
fetch_proposicoes()
}
\value{
Dataframe contendo todas as informações das proposições
}
\description{
Recupera da API todas as proposições disponíveis na base de dados da Câmara dos Deputados.
}
\examples{
proposicoes <- fetch_proposicoes()

}
