% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposicoes.R
\name{fetch_relacionadas}
\alias{fetch_relacionadas}
\alias{.fetch_relacionadas_camara}
\title{Fetches all propositions related to a proposition}
\usage{
fetch_relacionadas(casa, id_casa)

.fetch_relacionadas_camara(id_prop)
}
\arguments{
\item{casa}{senado or camara}

\item{id_casa}{Proposition's ID}

\item{id_prop}{Proposition's ID}
}
\value{
Dataframe containing all the related propositions.

Dataframe containing all the related propositions.
}
\description{
Returns all propositions related to a proposition by its id.

Returns all propositions related to a proposition by its id.
}
\examples{
\dontrun{
relacionadas_pec241 <- fetch_relacionadas("camara",2088351)
}
}
\seealso{
\code{\link[rcongresso]{fetch_id_proposicao_camara}}

\code{\link[rcongresso]{fetch_id_proposicao_camara}}
}
