% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposicoes.R
\name{fetch_relacionadas_senado}
\alias{fetch_relacionadas_senado}
\title{Fetches all propositions related to a proposition}
\usage{
fetch_relacionadas_senado(id_prop)
}
\arguments{
\item{id_prop}{Proposition's ID}
}
\value{
Dataframe containing all the related propositions.
}
\description{
Returns all propositions related to a proposition by its id.
}
\examples{
relacionadas_senado <- fetch_relacionadas_senado(91341)
}
