% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tramitacoes.R
\name{fetch_related_requerimentos}
\alias{fetch_related_requerimentos}
\title{Fetch related requirements}
\usage{
fetch_related_requerimentos(id, mark_deferimento = TRUE)
}
\arguments{
\item{id}{Proposition's id}

\item{mark_deferimento}{valor default true}
}
\value{
Dataframe
}
\description{
Retorns a dataframe with data from related requirements of a proposition
}
