% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requerimentos.R
\name{fetch_related_requerimentos_camara}
\alias{fetch_related_requerimentos_camara}
\title{Fetch related requerimentos in Camara}
\usage{
fetch_related_requerimentos_camara(prop_id, mark_deferimento = FALSE)
}
\arguments{
\item{prop_id}{ID of a Proposicao}

\item{mark_deferimento}{whether to retrieve status of requerimento}
}
\value{
Dataframe
}
\description{
Returns a dataframe with data from requerimentos related to a given proposition in Camara
}
