% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatorias.R
\name{fetch_relatorias}
\alias{fetch_relatorias}
\title{Fetches proposição's relatoria historic}
\usage{
fetch_relatorias(proposicao_id, casa, last_n = NULL)
}
\arguments{
\item{proposicao_id}{Proposição's ID}

\item{casa}{senado or câmara}

\item{last_n}{Fetch last N relatores}
}
\value{
Dataframe containing detailed information about relatorias of a proposição
}
\description{
Returns a dataframe containing the relatoria's historic, with date and relator's name.
}
\details{
Senado contains details about the end of each relatoria and camara contains initials of local 
from relatoria
}
\examples{
fetch_relatorias(91341, 'senado', 3)
}
