% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senadores.R
\name{fetch_senador}
\alias{fetch_senador}
\title{Fetches senator's info}
\usage{
fetch_senador(id = NULL)
}
\arguments{
\item{id}{Senator's id}
}
\value{
Dataframe containing the senator's info.
}
\description{
Returns its info, as name, gender, uf, situation, etc.
}
\examples{
senator <- fetch_senador(5573)
}
