% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senadores.R
\name{fetch_senadores}
\alias{fetch_senadores}
\title{Fetches info of the senators}
\usage{
fetch_senadores(legis_initial, legis_final)
}
\arguments{
\item{legis_initial}{Initial legisture}

\item{legis_final}{Final legisture}
}
\value{
Dataframe containing the senators info.
}
\description{
Returns its id, name, uf, gender.
}
\examples{
senators <- fetch_senadores(40, 60)
}
