% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposicoes.R, R/sessoes.R
\name{fetch_tramitacao}
\alias{fetch_tramitacao}
\alias{fetch_sessoes}
\title{Fetches the tramitation of a proposition}
\usage{
fetch_tramitacao(id_prop, casa)

fetch_sessoes(id_prop, casa)
}
\arguments{
\item{id_prop}{Proposition's ID}

\item{casa}{Proposition's casa}

\item{id_prop}{Proposition's ID}

\item{casa}{Proposition's casa}
}
\value{
Dataframe containing all the tramitation.

Dataframe containing all the sessions.
}
\description{
Returns the tramitation of a proposition by its id.

Returns all the sessions when a proposition was/is going to be discussed in a Congress House.
}
\examples{
tramitacao_pec241 <- fetch_sessoes(2088351,'camara')
tramitacao_pls229 <- fetch_sessoes(91341,'senado')
sessoes_pec_241_2016 <- fetch_sessoes(2088351,'camara')
sessoes_lei_qualidade_fiscal <- fetch_sessoes(91341,'senado')
}
