% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessoes.R
\name{fetch_sessoes_camara}
\alias{fetch_sessoes_camara}
\title{Fetches all the sessions when a proposition was/is going to be discussed in the Chamber of Deputies}
\usage{
fetch_sessoes_camara(id_prop)
}
\arguments{
\item{id_prop}{Proposition's ID}
}
\value{
Dataframe containing all the sessions.
}
\description{
Returns all the sessions when a proposition was/is going to be discussed in the Chamber of Deputies.
}
\examples{
sessoes_pec_241_2016 <- fetch_sessoes_camara(2088351)
}
