% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessoes.R
\name{fetch_sessoes_senado}
\alias{fetch_sessoes_senado}
\title{Fetches all the sessions when a proposition was/is going to be discussed in the Senate}
\usage{
fetch_sessoes_senado(id_prop)
}
\arguments{
\item{id_prop}{Proposition's ID}
}
\value{
Dataframe containing all the sessions.
}
\description{
Returns all the sessions when a proposition was/is going to be discussed in the Senate.
}
\examples{
sessoes_lei_qualidade_fiscal <- fetch_sessoes_senado(91341)
}
