% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposicoes.R
\name{fetch_status_proposicao}
\alias{fetch_status_proposicao}
\title{Recupera da API o status da proposição passada, incluindo sequência, órgão,
informações sobre a tramitação e data/hora da proposição.}
\usage{
fetch_status_proposicao(id_prop)
}
\arguments{
\item{id_prop}{ID da proposição}
}
\value{
Dataframe contendo informações sobre o status da proposição
}
\description{
Recupera da API o status da proposição passada, incluindo sequência, órgão,
informações sobre a tramitação e data/hora da proposição.
}
\examples{
pec241_status <- fetch_status_proposicao(2088351)

}
