% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposicoes.R
\name{fetch_temas_camara}
\alias{fetch_temas_camara}
\title{Fetch the proposition's themes}
\usage{
fetch_temas_camara(prop_id)
}
\arguments{
\item{prop_id}{Proposition's ID}
}
\value{
A dataframe containing the themes of a proposition
}
\description{
Returns a dataframe containing the themes of a proposition
}
\examples{
fetch_temas_camara(2121442)
}
