% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposicoes.R
\name{fetch_temas_proposicao}
\alias{fetch_temas_proposicao}
\title{Fetch the proposition's themes}
\usage{
fetch_temas_proposicao(proposicao_id, casa)
}
\arguments{
\item{proposicao_id}{Proposition's ID}

\item{casa}{senado or camara}
}
\value{
A dataframe containing the themes of a proposition
}
\description{
Returns a dataframe containing the themes of a proposition
}
\examples{
fetch_temas_proposicao(2121442, "camara")
}
