% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposicoes.R
\name{fetch_textos_proposicao}
\alias{fetch_textos_proposicao}
\title{Fetches all propositions related to a proposition}
\usage{
fetch_textos_proposicao(id)
}
\arguments{
\item{id}{Proposition's ID}
}
\value{
Dataframe containing all the related propositions.
}
\description{
Returns all propositions related to a proposition by its id.
}
\examples{
relacionadas_texto <- fetch_textos_proposicao(129808)
}
