% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposicoes.R
\name{fetch_tipo_proposicao}
\alias{fetch_tipo_proposicao}
\title{Dado um ID do tipo de proposição, retorna seu tipo.}
\usage{
fetch_tipo_proposicao(id_tipo_prop)
}
\arguments{
\item{id_tipo_prop}{ID do tipo da proposição}
}
\value{
Dataframe contendo o tipo da proposição
}
\description{
Dado um ID do tipo de proposição, retorna seu tipo.
}
\examples{
tipo_prop129 <- fetch_tipo_proposicao(129)

}
