% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposicoes.R
\name{fetch_tipos_proposicao}
\alias{fetch_tipos_proposicao}
\title{Fetches all the proposition types.}
\usage{
fetch_tipos_proposicao()
}
\value{
Proposition types
}
\description{
Fetches all the proposition types.
}
