% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tramitacoes.R
\name{fetch_tramitacao}
\alias{fetch_tramitacao}
\title{Fetches the tramitation of a proposition}
\usage{
fetch_tramitacao(id_prop, casa)
}
\arguments{
\item{id_prop}{Proposition's ID}

\item{casa}{Proposition's casa}
}
\value{
Dataframe containing all the tramitation.
}
\description{
Returns the tramitation of a proposition by its id.
}
\examples{
tramitacao_pec241 <- fetch_sessoes(2088351,'camara')
tramitacao_pls229 <- fetch_sessoes(91341,'senado')
}
