% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tramitacoes.R
\name{fetch_tramitacao_camara}
\alias{fetch_tramitacao_camara}
\title{Fetches the tramitation of a proposition in the Chamber of Deputies}
\usage{
fetch_tramitacao_camara(id_prop)
}
\arguments{
\item{id_prop}{Proposition's ID}
}
\value{
Dataframe containing all the tramitation.
}
\description{
Returns the tramitation of a proposition by its id.
}
\examples{
tramitacao_pec241 <- fetch_tramitacao_camara(2088351)
}
\seealso{
\code{\link[rcongresso]{fetch_id_proposicao}}
}
