% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tramitacoes.R
\name{fetch_tramitacao_senado}
\alias{fetch_tramitacao_senado}
\title{Fetches the tramitation of a proposition in the Senate}
\usage{
fetch_tramitacao_senado(id_prop)
}
\arguments{
\item{id_prop}{Proposition's ID}
}
\value{
Dataframe containing all the tramitation.
}
\description{
Returns the tramitation of a proposition by its id.
}
\examples{
tramitacao_pls229 <- fetch_tramitacao_senado(91341)
}
\seealso{
\code{\link[rcongresso]{fetch_id_proposicao}}
}
