% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deputados.R
\name{fetch_ultimo_status_deputado}
\alias{fetch_ultimo_status_deputado}
\title{Fetches the last information about the deputy's mandate}
\usage{
fetch_ultimo_status_deputado(dep_id)
}
\arguments{
\item{dep_id}{deputy's ID}
}
\value{
Dataframe containing details about deputy's party, electoral name and url photo
}
\description{
Fetches information such as party and electoral name.
}
\examples{
abel_mesquita_ultimo_estado <- fetch_ultimo_status_deputado(dep_id = 178957)
}
