% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votacoes.R
\name{fetch_votacao}
\alias{fetch_votacao}
\title{Recupera da API detalhes sobre uma votação específica.}
\usage{
fetch_votacao(id_votacao)
}
\arguments{
\item{id_votacao}{ID da votação}
}
\value{
Dataframe contendo os detalhes de uma votação, incluindo título, a hora de início da votação,
hora de término da votação, placar e sua aprovação.
}
\description{
Recupera da API detalhes sobre uma votação específica.
}
\examples{
votacao_segundoturno_pec241 <- fetch_votacao(7252)

}
