% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposicoes.R
\name{fetch_votacoes}
\alias{fetch_votacoes}
\title{Recupera da API todas as votações pelas quais uma proposição já passou.}
\usage{
fetch_votacoes(id_props)
}
\arguments{
\item{id_prop}{ID da proposição}
}
\value{
Dataframe contendo as várias votações pelas quais uma proposição passou.
}
\description{
Recupera da API todas as votações pelas quais uma proposição já passou.
}
\examples{
votacoes_pec241 <- fetch_votacoes(2088351)

}
