% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R, R/votacoes.R
\name{.rename_votacoes_df}
\alias{.rename_votacoes_df}
\alias{fetch_votacoes_senado}
\title{Renames the cols of the bill's voting on Senate}
\usage{
.rename_votacoes_df(df)

fetch_votacoes_senado(proposicao_id)
}
\arguments{
\item{df}{Dataframe}

\item{proposicao_id}{Proposition Id}
}
\value{
Dataframe containing the renamed strings.

Dataframe containing details about a voting on Senate
}
\description{
Renames each item from vector with the pattern: split by underscore and lowercase

Fetches details about a voting on Senate.
Ao fim, a função retira todos as colunas que tenham tipo lista para uniformizar o dataframe.
}
\examples{
fetch_votacoes_senado(91341)
}
