% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votacoes.R
\name{fetch_votos}
\alias{fetch_votos}
\title{Fetches the votes from a voting.}
\usage{
fetch_votos(id_votacao)
}
\arguments{
\item{id_votacao}{ID da votação}
}
\value{
Dataframe containing all the deputy's vote on this voting.
}
\description{
Fetches the votes from a voting.
}
\examples{
votos_segundoturno_pec241 <- fetch_votos(7252)

}
