% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventos.R
\name{get_next_eventos_camara}
\alias{get_next_eventos_camara}
\title{Recupera os próximos eventos (sessões/reuniões) de uma proposição na Câmara}
\usage{
get_next_eventos_camara(prop_id)
}
\arguments{
\item{prop_id}{ID da proposição}
}
\value{
Dataframe contendo o timestamp, o local e a descrição do evento.
}
\description{
Retorna um dataframe contendo o timestamp, o local e a descrição do evento
}
\examples{
get_next_eventos_camara(2121442)
}
