% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votacoes.R
\name{get_votos_partidos}
\alias{get_votos_partidos}
\title{Dada uma votação, retorna o posicionamento de cada partido naquela votação.}
\usage{
get_votos_partidos(votacao)
}
\arguments{
\item{votacao}{Dataframe contendo informações sobre a votação}
}
\description{
Dada uma votação, retorna o posicionamento de cada partido naquela votação.
}
\examples{
pos_partidos <- get_votos_partidos(7252)

}
