% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcongresso.R
\docType{package}
\name{rcongresso}
\alias{rcongresso}
\alias{rcongresso-package}
\title{rcongresso: Pacote R para acessar dados do congresso nacional.}
\description{
O pacote dá acesso à várias funções que são divididas em categorias: Proposições, deputados e votações.
Veja mais detalhes nas funções ou consultando o site da câmara na referência abaixo.
}
\references{
\url{https://dadosabertos.camara.leg.br/swagger/api.html}
}
