% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_table_to_underscore}
\alias{rename_table_to_underscore}
\title{Renomeia as colunas do dataframe passado para o formato underscore}
\usage{
rename_table_to_underscore(df)
}
\arguments{
\item{df}{Dataframe do Senado}
}
\value{
Dataframe com as colunas renomeadas
}
\description{
Renomeia as colunas do dataframe usando o padrão
de underscore e letras minúsculas
}
