% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_function_with_delay}
\alias{run_function_with_delay}
\title{Runs a given function after a specified delay}
\usage{
run_function_with_delay(delay, fun)
}
\arguments{
\item{delay}{delay to sleep before running function}

\item{fun}{function (with/out parameters or not) to be run after delay}
}
\value{
the return value of the function
}
\description{
Runs a given function after a specified delay
}
