% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposicoes.R
\name{scrap_autores_from_website}
\alias{scrap_autores_from_website}
\title{Scraps autores da proposição from website}
\usage{
scrap_autores_from_website(id_prop)
}
\arguments{
\item{id_prop}{proposição's ID}
}
\value{
String with authors names separated by comma
}
\description{
Return the author(s) name(s)
}
