% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposicoes.R
\name{scrap_senado_documentos}
\alias{scrap_senado_documentos}
\title{Scraps data from documents of a bill from website}
\usage{
scrap_senado_documentos(id_prop, filter_texto_materia = F)
}
\arguments{
\item{id_prop}{proposition's ID}

\item{filter_texto_materia}{Bool flag to filter the initial texts}
}
\value{
dataframe
}
\description{
Return the document's data
}
