% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_underscore}
\alias{to_underscore}
\title{Renomeia um vetor com o padrão de underscores e minúsculas}
\usage{
to_underscore(x)
}
\arguments{
\item{x}{Vetor de strings}
}
\value{
Vetor contendo as strings renomeadas.
}
\description{
Renomeia cada item do vetor com o padrão: separado por underscore e letras minúsculas
}
\examples{
to_underscore(c("testName", "TESTNAME"))
}
