% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votacoes.R
\name{ultima_votacao}
\alias{ultima_votacao}
\title{Recebe um dataframe de votações e seleciona a última votação de cada proposição.}
\usage{
ultima_votacao(votacoes)
}
\arguments{
\item{votacoes}{Dataframe contendo as votações a serem filtradas}
}
\value{
Dataframe contendo apenas as últimas votações relacionadas às proposições
}
\description{
Recebe um dataframe de votações e seleciona a última votação de cada proposição.
}
\examples{
votacoes_pec241 <- fetch_votacoes(2088351)
ultima_votacao <- ultima_votacao(votacoes_pec241)

}
