% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmr.R
\name{constroi_dataframe}
\alias{constroi_dataframe}
\title{Função que constroi o dataframe modelo utilizado para as análises realizadas pela plataforma "Quem me representa?"}
\usage{
constroi_dataframe(proposicao, votacao)
}
\arguments{
\item{proposicao}{Um dataframe contendo proposições recuperado pelo método fetch_proposicao()}

\item{votacao}{Uma dataframe contendo votações recuperado pelo método fetch_votacao()}
}
\value{
Dataframe contendo o resultado do join desses dataframes com os votos ocorridos nas votações
passadas como parâmetro.
}
\description{
Função que constroi o dataframe modelo utilizado para as análises realizadas pela plataforma "Quem me representa?"
}
\examples{
pec241 <- fetch_proposicao(2088351)
votacao_segundoturno_pec241 <- fetch_votacao(7252)
dataframe_pec241 <- constroi_dataframe(pec241, votacao_segundoturno_pec241)

}
