% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.fetch_using_id}
\alias{.fetch_using_id}
\title{Fetches details from a proposition.}
\usage{
.fetch_using_id(id, API_path, asList = FALSE)
}
\arguments{
\item{id}{Proposition's ID}

\item{API_path}{API path}

\item{asList}{If return should be a list or a dataframe}
}
\value{
Dataframe containing information about the proposition.
}
\description{
Fetches details from a proposition.
}
\examples{
pec241 <- rcongresso:::.fetch_using_id(2088351, API_path = "/api/v2/proposicoes")

}
