% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.fetch_using_queries}
\alias{.fetch_using_queries}
\title{Fetches a proposition using a list of queries}
\usage{
.fetch_using_queries(parametros, API_path, asList = FALSE)
}
\arguments{
\item{parametros}{queries used on the search}

\item{API_path}{API path}

\item{asList}{If return should be a list or a dataframe}
}
\value{
Dataframe containing information about the proposition.
}
\description{
Fetches a proposition using a list of queries
}
\examples{
pec241 <- rcongresso:::.fetch_using_queries(
   parametros = list(id = "2088351"),
   API_path = "/api/v2/proposicoes"
)

}
