% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_json}
\alias{.get_json}
\title{Extracts the JSON data from an HTTP response}
\usage{
.get_json(response)
}
\arguments{
\item{response}{The HTTP response}
}
\value{
The json
}
\description{
Extracts the JSON data from an HTTP response
}
