% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agenda.R
\name{fetch_agenda_camara}
\alias{fetch_agenda_camara}
\title{fetch_agenda_camara('2018-07-03', '2018-07-10')}
\usage{
fetch_agenda_camara(initial_date, end_date)
}
\description{
fetch_agenda_camara('2018-07-03', '2018-07-10')
}
