% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deputados.R
\name{fetch_all_deputados}
\alias{fetch_all_deputados}
\alias{fetch_ids_deputados}
\title{Fetches details abaout all deputys}
\usage{
fetch_all_deputados(ids_dep)

fetch_ids_deputados(legislatura_base = .LEGISLATURA_INICIAL)
}
\arguments{
\item{ids_dep}{Dataframe containing all deputys IDs}

\item{legislatura_base}{Legislatura inicial para retornar os deputados}
}
\value{
Dataframe containing details about the deputy's

Dataframe containing all deputys IDs
}
\description{
Fetches details about deputys from the 40º legislature to the current

Fetches all deputys IDs from the given legislature to the current
}
