% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposicoes.R
\name{fetch_apensadas_camara}
\alias{fetch_apensadas_camara}
\title{Recupera as proposições apensadas}
\usage{
fetch_apensadas_camara(prop_id)
}
\arguments{
\item{prop_id}{ID da proposição}
}
\value{
Ventor contendo os IDs das proposições apensadas
}
\description{
Retorna os IDs das proposições apensadas a uma determinada proposição
}
\examples{
fetch_apensadas(2121442)
}
