% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comissoes.R
\name{fetch_composicao_comissoes_senado}
\alias{fetch_composicao_comissoes_senado}
\title{Retorna a composição da comissão do senado}
\usage{
fetch_composicao_comissoes_senado(sigla)
}
\arguments{
\item{sigla}{Sigla da comissão do Senado}
}
\value{
Dataframes
}
\description{
Retorna dataframe com os dados dos membros de uma comissão do Senado
}
