% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deputados.R
\name{fetch_deputado}
\alias{fetch_deputado}
\title{Retrieves details about a deputy}
\usage{
fetch_deputado(id = NULL, nome = NULL, idLegislatura = NULL,
  siglaUf = NULL, siglaPartido = NULL, siglaSexo = NULL, itens = NULL,
  ordenarPor = NULL)
}
\arguments{
\item{id}{Deputy's ID}

\item{nome}{Deputy's name (partially or complete)}

\item{idLegislatura}{Id of one or more mandates of the deputy}

\item{siglaUf}{State's abbreviation of the deputy}

\item{siglaPartido}{Party's abbreviation of the deputy}

\item{siglaSexo}{Gender's abbreviation}

\item{itens}{Items quantity. '-1' returns all current deputies at the request moment}

\item{ordenarPor}{Element that will be used to sort the returned list}
}
\value{
Dataframe containing details about the deputy.
}
\description{
ID, name, birth date, birth city among other informations are returned.
}
\examples{
abel_mesquita_info <- fetch_deputado(id = 178957)
}
