% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emendas.R
\name{fetch_emendas}
\alias{fetch_emendas}
\title{Returns emendas of a proposição from Congresso}
\usage{
fetch_emendas(id, casa)
}
\arguments{
\item{bill_id}{Proposição's ID}
}
\value{
Dataframe with informations about emendas of a proposição.
}
\description{
Fetchs a dataframe with emendas's data of a proposição from Congresso.
}
\examples{
fetch_emendas(91341,'senado')
}
