% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emendas.R
\name{fetch_emendas_camara}
\alias{fetch_emendas_camara}
\title{Returns emendas of a proposição from Câmara.}
\usage{
fetch_emendas_camara(bill_id)
}
\arguments{
\item{bill_id}{Proposição's ID from Câmara.}
}
\value{
Dataframe with informations about emendas of a proposição from Câmara.
}
\description{
Fetchs a dataframe with emendas's data of a proposição from Câmara.
}
