% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emendas.R
\name{fetch_emendas_camara}
\alias{fetch_emendas_camara}
\title{Fetches proposition's emendas}
\usage{
fetch_emendas_camara(sigla = NULL, numero = NULL, ano = NULL)
}
\arguments{
\item{sigla}{Proposition type (i.e., PEC, PL, PDC)}

\item{numero}{Proposition number}

\item{ano}{Proposition year}
}
\value{
A dataframe containing details about the emendas of the proposition
}
\description{
Fetches a dataframe containing the emendas of the proposition
}
\examples{
fetch_emendas_camara('pl', 6726, 2016)
}
