% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgaos.R
\name{fetch_orgaos_camara}
\alias{fetch_orgaos_camara}
\title{Fetches data about all Câmara's órgãos}
\usage{
fetch_orgaos_camara()
}
\value{
A dataframe
}
\description{
Fetches a dataframe containing information about all Câmara's órgãos
}
