% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votacoes.R
\name{fetch_orientacoes}
\alias{fetch_orientacoes}
\title{Recupera da API as orientações das bancadas em uma determinada votação.}
\usage{
fetch_orientacoes(id_votacao)
}
\arguments{
\item{id_votacao}{ID da votação}
}
\value{
Dataframe contendo as orientações das bancadas na votação especificada
}
\description{
Recupera da API as orientações das bancadas em uma determinada votação.
}
\examples{
orientacoesbancada_segundoturno_pec241 <- fetch_orientacoes(7252)

}
