% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votacoes.R
\name{fetch_orientacoes}
\alias{fetch_orientacoes}
\title{Fetches the positions of a group on a voting}
\usage{
fetch_orientacoes(id_votacao = NULL)
}
\arguments{
\item{id_votacao}{Voting's ID}
}
\value{
Dataframe containing the groups and their votes
}
\description{
Fetch how the groups in the chamber of deputies
instructed their members to vote on a given voting.
}
\examples{
orientacoes_pec241 <- fetch_orientacoes(7252)
}
\seealso{
\code{\link[rcongresso]{get_votos_partidos}}
}
