% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposicoes.R
\name{fetch_proposicao}
\alias{fetch_proposicao}
\title{Fetches a proposition in a specific house}
\usage{
fetch_proposicao(proposicao_id = NULL, casa)
}
\arguments{
\item{proposicao_id}{Proposition's ID}

\item{casa}{'camara' or 'senado'}
}
\value{
Dataframe containing all the info about the proposition
}
\description{
Returns the proposition info
}
\examples{
prop_pls229 <- fetch_proposicao(91341, 'senado')
}
