% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposicoes.R
\name{fetch_proposicao}
\alias{fetch_proposicao}
\title{Fetches details from a proposition.}
\usage{
fetch_proposicao(id_prop)
}
\arguments{
\item{id_prop}{Proposition's ID}
}
\value{
Dataframe containing information about the proposition.
}
\description{
Fetches details from a proposition.
}
\examples{
pec241 <- fetch_proposicao(2088351)

}
