% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposicoes.R
\name{fetch_proposicao}
\alias{fetch_proposicao}
\alias{fetch_proposicao_senado}
\title{Fetches a proposition in a specific house}
\usage{
fetch_proposicao(proposicao_id = NULL, casa)

fetch_proposicao_senado(id = NULL)
}
\arguments{
\item{casa}{'camara' or 'senado'}

\item{id}{Proposition's ID}

\item{id}{Proposition's ID}
}
\value{
Dataframe containing all the info about the proposition;

Dataframe containing all the info about the proposition;
}
\description{
Returns the proposition info

Returns the proposition info
}
\examples{
prop_pls229 <- fetch_proposicao_senado(91341, 'senado')
prop_pls229 <- fetch_proposicao_senado(91341)
}
