% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requerimentos.R
\name{fetch_related_requerimentos}
\alias{fetch_related_requerimentos}
\title{Fetch related requerimentos}
\usage{
fetch_related_requerimentos(prop_id, casa)
}
\arguments{
\item{prop_id}{ID of a Proposicao}

\item{casa}{camara or senado}
}
\value{
Dataframe
}
\description{
Returns a dataframe with data from requerimentos related to a given proposition
}
