% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requerimentos.R
\name{fetch_related_requerimentos_senado}
\alias{fetch_related_requerimentos_senado}
\title{Fetch related requerimentos in Senado}
\usage{
fetch_related_requerimentos_senado(prop_id, normalized = FALSE)
}
\arguments{
\item{prop_id}{ID of a Proposicao}

\item{normalized}{True or False}
}
\value{
Dataframe
}
\description{
Returns a dataframe with data from requerimentos related to a given proposition in Senado
}
