% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessoes.R
\name{fetch_sessoes}
\alias{fetch_sessoes}
\title{Fetches all the sessions when a proposition was/is going to be discussed in a Congress House}
\usage{
fetch_sessoes(id_prop, casa)
}
\arguments{
\item{id_prop}{Proposition's ID}

\item{casa}{Proposition's casa}
}
\value{
Dataframe containing all the sessions.
}
\description{
Returns all the sessions when a proposition was/is going to be discussed in a Congress House.
}
\examples{
sessoes_pec_241_2016 <- fetch_sessoes(2088351,'camara')
sessoes_lei_qualidade_fiscal <- fetch_sessoes(91341,'senado')
}
