% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposicoes.R
\name{fetch_status_proposicao}
\alias{fetch_status_proposicao}
\title{Recovers the proposition status in the parlament}
\usage{
fetch_status_proposicao(id_prop)
}
\arguments{
\item{id_prop}{Proposition's ID}
}
\value{
Dataframe containing the proposition's status
}
\description{
Recovers the proposition status including: sequence, organ
and date info about its processing in the parlament.
}
\examples{
pec241_status <- fetch_status_proposicao(2088351)
}
\seealso{
\code{\link[rcongresso]{fetch_proposicao}}, \code{\link[rcongresso]{fetch_id_proposicao}}
}
