% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposicoes.R
\name{fetch_tipo_proposicao}
\alias{fetch_tipo_proposicao}
\title{Fetches the type of the proposition from its id.}
\usage{
fetch_tipo_proposicao(id_tipo_prop)
}
\arguments{
\item{id_tipo_prop}{Proposition's type ID}
}
\value{
Dataframe containing the proposition's type info.
}
\description{
Fetches the type of the proposition from its id.
}
\examples{
tipo_prop129 <- fetch_tipo_proposicao(129)

}
