% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposicoes.R
\name{fetch_tipos_proposicao}
\alias{fetch_tipos_proposicao}
\title{Recupera da API os tipos de proposição disponíveis.}
\usage{
fetch_tipos_proposicao()
}
\value{
Tipos de proposição
}
\description{
Recupera da API os tipos de proposição disponíveis.
}
\examples{
tipos_proposicao <- fetch_tipos_proposicao()

}
