% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votacoes.R
\name{fetch_votacao}
\alias{fetch_votacao}
\title{Fetches details about a voting.}
\usage{
fetch_votacao(id_votacao)
}
\arguments{
\item{id_votacao}{Voting's ID}
}
\value{
Dataframe containing details about a voting, including tittle,
 start voting time, finish voting time, result and approval.
}
\description{
Fetches details about a voting.
}
\examples{
segundoturno_pec241 <- fetch_votacao(7252)

}
