% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposicoes.R
\name{fetch_votacoes}
\alias{fetch_votacoes}
\title{Fetches all the votings which a proposition went through.}
\usage{
fetch_votacoes(id_prop)
}
\arguments{
\item{id_prop}{Proposition's ID}
}
\value{
Dataframe containing all the votings.
}
\description{
Fetches all the votings which a proposition went through.
}
\examples{
votacoes_pec241 <- fetch_votacoes(2088351)

}
