% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votacoes.R
\name{fetch_votos}
\alias{fetch_votos}
\title{Recupera os votos referentes àquela votação específica.}
\usage{
fetch_votos(id_votacao)
}
\arguments{
\item{id_votacao}{ID da votação}
}
\value{
Dataframe contendo o posicionamento de cada deputado, além de informações sobre estes.
}
\description{
Recupera os votos referentes àquela votação específica.
}
\examples{
votos_segundoturno_pec241 <- fetch_votos(7252)

}
