% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votacoes.R
\name{fetch_votos}
\alias{fetch_votos}
\title{Fetches individual votes from a voting.}
\usage{
fetch_votos(id_votacao)
}
\arguments{
\item{id_votacao}{ID da votação}
}
\value{
Dataframe containing votes for each deputy on this voting.
}
\description{
Fetches individual votes from a voting.
}
\examples{
votos_pec241 <- fetch_votos(7252)

}
