% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventos.R
\name{get_next_eventos_camara}
\alias{get_next_eventos_camara}
\title{Get the next events about a proposition in the Câmara}
\usage{
get_next_eventos_camara(prop_id)
}
\arguments{
\item{prop_id}{Proposition ID}
}
\value{
Dataframe
}
\description{
Returns a dataframe with timestamp, origen and description about each event
}
\examples{
get_next_eventos_camara(2121442)
}
