% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votacoes.R
\name{get_votos_partidos}
\alias{get_votos_partidos}
\title{Fetches the positions by party on a voting}
\usage{
get_votos_partidos(votacao = NULL)
}
\arguments{
\item{votacao}{Voting's id}
}
\value{
Dataframe with each party's orientation for its deputies.
}
\description{
Fetch how parties in the chamber of deputies
instructed their members to vote on a given voting.
}
\examples{
pos_partidos <- get_votos_partidos(7252)
}
\seealso{
\code{\link[rcongresso]{fetch_orientacoes}}
}
