% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votacoes.R
\name{get_votos_partidos}
\alias{get_votos_partidos}
\title{Fetch how parties in the chamber of deputies
instructed their members to vote on a given voting.}
\usage{
get_votos_partidos(votacao)
}
\arguments{
\item{votacao}{Voting id}
}
\value{
Dataframe with each party's orientation for its deputies.
}
\description{
Fetch how parties in the chamber of deputies
instructed their members to vote on a given voting.
}
\examples{
pos_partidos <- get_votos_partidos(7252)

}
