% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcongresso.R
\docType{package}
\name{rcongresso}
\alias{rcongresso}
\alias{rcongresso-package}
\title{rcongresso: R Package to access Brazil's national congress data based on its RESTful API.}
\description{
The package wraps API endpoints for propositions, deputies, votings and votes.
}
\details{
See more on references.
}
\references{
\url{https://dadosabertos.camara.leg.br/swagger/api.html}
}
