% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_table_to_underscore}
\alias{rename_table_to_underscore}
\title{Changes the column names of the input dataframe to underscore}
\usage{
rename_table_to_underscore(df)
}
\arguments{
\item{df}{input dataframe}
}
\value{
dataframe with underscore column names
}
\description{
Changes the column names of the input dataframe to underscore
}
